Add namespace cache to the Pelican filesystem
The namespace cache keeps track of the endpoints associated with a
given prefix of the federation namespace.  It is a "ttlcache", meaning
the data will be periodically refreshed from the dirctor as it
becomes stale.

The cache can be updated to remove problematic endpoints; in the future,
this will allow us to intercept I/O failures and mark certain endpoints
as unusable.